+++ python/coverage-tests/configuration/playwright.js
@@ -6,9 +6,19 @@ let report_package_name = {
 }
 module.exports = {
     name: report_package_name[process.env.RELEASING_PACKAGE] || "eyes_selenium_python",
+    env: {
+        AUTOMATION_FRAMEWORK: "playwright"
+    },
+    emitOnly: test => {
+        let env = test.env || {}
+        let features = test.features || []
+        return !env.device &&
+            !["ie-11", "edge-18", "safari-11", "safari-12"].includes(env.browser) &&
+            !features.includes("webdriver")
+    },
     emitter: `${dir}/python/emitter.js`,
     overrides: [`${dir}/js/overrides.js`, `${dir}/python/overrides.js`],
-    template: `${dir}/python/template.hbs`,
+    template: `${dir}/python/playwright-template.hbs`,
     tests: `${dir}/coverage-tests.js`,
     ext: ".py",
     outPath: "./test/coverage/generic",