+++ python/coverage-tests/test/__init__.py
@@ -1,1 +1,1 @@
-from .util import get_dom, get_test_info, getNodesByAttribute
+from .util import By, get_dom, get_test_info, getNodesByAttribute