+++ python/coverage-tests/test/browsers.py
@@ -1,19 +1,23 @@
 import os
-import time
 
 import pytest
-import selenium
 from pkg_resources import parse_version
-from selenium import webdriver
-from webdriver_manager.chrome import ChromeDriverManager
-from webdriver_manager.firefox import GeckoDriverManager
 
 from . import sauce
 
-LEGACY_SELENIUM = parse_version(selenium.__version__) < parse_version("4")
-# Download driver during module import to avoid racy downloads by xdist workers
-GECKO_DRIVER = GeckoDriverManager().install()
-CHROME_DRIVER = ChromeDriverManager().install()
+try:
+    import selenium
+    from selenium import webdriver
+    from webdriver_manager.chrome import ChromeDriverManager
+    from webdriver_manager.firefox import GeckoDriverManager
+
+    LEGACY_SELENIUM = parse_version(selenium.__version__) < parse_version("4")
+    # Download driver during module import to avoid racy downloads by xdist workers
+    GECKO_DRIVER = GeckoDriverManager().install()
+    CHROME_DRIVER = ChromeDriverManager().install()
+except ImportError:
+    # No selenium installed, browser fixtures will not work
+    pass
 
 
 @pytest.fixture(scope="function")