+++ python/eyes_images/applitools/common/test_results.py
@@ -151,7 +151,9 @@ class TestResults(object):
         from applitools.selenium.command_executor import CommandExecutor
         from applitools.selenium.eyes import EyesRunner
 
-        cmd = CommandExecutor.get_instance(EyesRunner.BASE_AGENT_ID, __version__)
+        cmd = CommandExecutor.get_instance(
+            EyesRunner.Protocol, EyesRunner.BASE_AGENT_ID, __version__
+        )
         cmd.core_delete_test(self)