+++ python/tests/functional/eyes_selenium/test_usdk_commands.py
@@ -5,15 +5,17 @@ from os import getcwd
 from mock import ANY
 from pytest import raises
 
+from applitools.common.connection import USDKConnection
 from applitools.common.errors import USDKFailure
 from applitools.common.selenium import Configuration
 from applitools.selenium.command_executor import CommandExecutor, ManagerType
-from applitools.selenium.connection import USDKConnection
 from applitools.selenium.fluent import SeleniumCheckSettings
+from applitools.selenium.object_registry import SeleniumWebdriverObjectRegistry
+from applitools.selenium.protocol import SeleniumWebDriver
 
 
 def test_usdk_commands_make_manager():
-    commands = CommandExecutor(USDKConnection.create())
+    commands = CommandExecutor(SeleniumWebDriver, USDKConnection.create())
     commands.make_core("sdk_name", "sdk_version", getcwd())
 
     mgr = commands.core_make_manager(ManagerType.CLASSIC)
@@ -24,19 +26,24 @@ def test_usdk_commands_make_manager():
 def test_usdk_commands_open_eyes(local_chrome_driver):
     configuration = Configuration().set_app_name("a").set_test_name("b")
     local_chrome_driver.get("https://demo.applitools.com")
-    commands = CommandExecutor(USDKConnection.create())
+    commands = CommandExecutor(SeleniumWebDriver, USDKConnection.create())
 
     commands.make_core("sdk_name", "sdk_version", getcwd())
 
     mgr = commands.core_make_manager(ManagerType.CLASSIC)
 
-    eyes = commands.manager_open_eyes(mgr, local_chrome_driver, config=configuration)
+    eyes = commands.manager_open_eyes(
+        SeleniumWebdriverObjectRegistry(),
+        mgr,
+        local_chrome_driver,
+        config=configuration,
+    )
 
     assert "applitools-ref-id" in eyes
 
 
 def test_usdk_commands_set_get_viewport_size(local_chrome_driver):
-    commands = CommandExecutor(USDKConnection.create())
+    commands = CommandExecutor(SeleniumWebDriver, USDKConnection.create())
     commands.make_core("sdk_name", "sdk_version", getcwd())
 
     commands.core_set_viewport_size(local_chrome_driver, {"width": 800, "height": 600})
@@ -52,14 +59,15 @@ def test_usdk_commands_open_close_eyes(local_chrome_driver):
         .set_test_name("USDK Commands open close")
         .set_user_test_id("42")
     )
-    commands = CommandExecutor(USDKConnection.create())
+    objreg = SeleniumWebdriverObjectRegistry()
+    commands = CommandExecutor(SeleniumWebDriver, USDKConnection.create())
     commands.make_core("sdk_name", "sdk_version", getcwd())
     mgr = commands.core_make_manager(ManagerType.CLASSIC)
-    eyes = commands.manager_open_eyes(mgr, local_chrome_driver, config=config)
+    eyes = commands.manager_open_eyes(objreg, mgr, local_chrome_driver, config=config)
 
     assert "applitools-ref-id" in mgr
 
-    commands.eyes_close(eyes, False, Configuration())
+    commands.eyes_close(objreg, eyes, False, Configuration())
     eyes_close_result = commands.eyes_get_results(eyes, False)
     test_result = eyes_close_result[0]
 
@@ -88,14 +96,15 @@ def test_usdk_commands_open_abort_eyes(local_chrome_driver):
         .set_test_name("USDK Commands open abort")
         .set_user_test_id("abc")
     )
-    commands = CommandExecutor(USDKConnection.create())
+    objreg = SeleniumWebdriverObjectRegistry()
+    commands = CommandExecutor(SeleniumWebDriver, USDKConnection.create())
     commands.make_core("sdk_name", "sdk_version", getcwd())
     mgr = commands.core_make_manager(ManagerType.CLASSIC)
-    eyes = commands.manager_open_eyes(mgr, local_chrome_driver, config=config)
+    eyes = commands.manager_open_eyes(objreg, mgr, local_chrome_driver, config=config)
 
     assert "applitools-ref-id" in mgr
 
-    commands.eyes_abort(eyes)
+    commands.eyes_abort(objreg, eyes)
     eyes_abort_result = commands.eyes_get_results(eyes, False)
 
     test_result = eyes_abort_result[0]
@@ -141,16 +150,17 @@ def test_usdk_commands_open_check_close_eyes(local_chrome_driver):
         .set_user_test_id("abc")
     )
 
-    commands = CommandExecutor(USDKConnection.create())
+    objreg = SeleniumWebdriverObjectRegistry()
+    commands = CommandExecutor(SeleniumWebDriver, USDKConnection.create())
     commands.make_core("sdk_name", "sdk_version", getcwd())
     mgr = commands.core_make_manager(ManagerType.CLASSIC)
-    eyes = commands.manager_open_eyes(mgr, local_chrome_driver, config=config)
+    eyes = commands.manager_open_eyes(objreg, mgr, local_chrome_driver, config=config)
 
     check_result = commands.eyes_check(
-        eyes, local_chrome_driver, SeleniumCheckSettings(), config
+        objreg, eyes, None, SeleniumCheckSettings(), config
     )
 
-    commands.eyes_close(eyes, False, Configuration())
+    commands.eyes_close(objreg, eyes, False, Configuration())
     eyes_close_result = commands.eyes_get_results(eyes, False)
     test_result = eyes_close_result[0]
 
@@ -175,9 +185,11 @@ def test_usdk_commands_open_check_close_eyes(local_chrome_driver):
 
 
 def test_usdk_commands_error_logging(caplog):
-    commands = CommandExecutor.get_instance("sdk_name", "sdk_version")
+    commands = CommandExecutor.get_instance(
+        SeleniumWebDriver, "sdk_name", "sdk_version"
+    )
 
     with raises(USDKFailure):
-        commands.manager_open_eyes({})
+        commands.manager_open_eyes(SeleniumWebdriverObjectRegistry(), {})
 
     assert "Re-raising an error received from SDK server: USDKFailure" in caplog.text