+++ python/tests/unit/eyes_selenium/test_ec_client.py
@@ -1,4 +1,4 @@
-from mock import call, patch
+from mock import ANY, call, patch
 from six import PY2
 
 from applitools.common import ProxySettings
@@ -13,12 +13,12 @@ def test_get_execution_cloud_url_no_args(monkeypatch):
 
         if PY2:  # work around __getitem__ bug in old mock
             assert c.mock_calls[:-1] == [
-                call("Core.makeECClient", {"settings": {}}),
+                call(ANY, "Core.makeECClient", {"settings": {}}),
             ]
             assert str(c.mock_calls[-1]) == "call().__getitem__(u'url')"
         else:
             assert c.mock_calls == [
-                call("Core.makeECClient", {"settings": {}}),
+                call(ANY, "Core.makeECClient", {"settings": {}}),
                 call().__getitem__("url"),
             ]
 
@@ -32,6 +32,7 @@ def test_get_execution_cloud_url_all_args(monkeypatch):
         if PY2:
             assert c.mock_calls[:-1] == [
                 call(
+                    ANY,
                     "Core.makeECClient",
                     {
                         "settings": {
@@ -49,6 +50,7 @@ def test_get_execution_cloud_url_all_args(monkeypatch):
         else:
             assert c.mock_calls == [
                 call(
+                    ANY,
                     "Core.makeECClient",
                     {
                         "settings": {