+++ python/tests/unit/eyes_selenium/test_target_path.py
@@ -13,6 +13,12 @@ from applitools.selenium.fluent.target_path import (
     ShadowDomLocator,
     TargetPath,
 )
+from applitools.selenium.object_registry import SeleniumWebdriverObjectRegistry
+
+
+@pytest.fixture
+def obj_reg():
+    return SeleniumWebdriverObjectRegistry()
 
 
 class DummyElement(object):
@@ -28,7 +34,7 @@ def test_target_path_region_by_css():
 
     assert type(path) is RegionLocator
     assert type(path.value) is ElementSelector
-    assert path.value.by is By.CSS_SELECTOR
+    assert path.value.by == By.CSS_SELECTOR
     assert path.value.selector == ".css"
     assert path.parent is None
 
@@ -38,7 +44,7 @@ def test_target_path_region_by_xpath():
 
     assert type(path) is RegionLocator
     assert type(path.value) is ElementSelector
-    assert path.value.by is By.XPATH
+    assert path.value.by == By.XPATH
     assert path.value.selector == "//x"
     assert path.parent is None
 
@@ -66,7 +72,7 @@ def test_target_path_frame_by_css_selector():
 
     assert type(path) is FrameLocator
     assert type(path.value) is ElementSelector
-    assert path.value.by is By.CSS_SELECTOR
+    assert path.value.by == By.CSS_SELECTOR
     assert path.value.selector == ".css"
     assert path.parent is None
 
@@ -76,7 +82,7 @@ def test_target_path_shadow_by_css():
 
     assert type(path) is ShadowDomLocator
     assert type(path.value) is ElementSelector
-    assert path.value.by is By.CSS_SELECTOR
+    assert path.value.by == By.CSS_SELECTOR
     assert path.value.selector == ".css"
     assert path.parent is None
 
@@ -86,7 +92,7 @@ def test_target_path_shadow_by_xpath():
 
     assert type(path) is ShadowDomLocator
     assert type(path.value) is ElementSelector
-    assert path.value.by is By.XPATH
+    assert path.value.by == By.XPATH
     assert path.value.selector == "//x"
     assert path.parent is None
 
@@ -123,7 +129,7 @@ def test_target_region_within_frame():
 
     assert type(path) is RegionLocator
     assert type(path.value) is ElementSelector
-    assert path.value.by is By.CSS_SELECTOR
+    assert path.value.by == By.CSS_SELECTOR
     assert path.value.selector == ".css"
     assert type(path.parent) is FrameLocator
     assert type(path.parent.value) is FrameSelector
@@ -248,26 +254,26 @@ def test_target_path_frame_region_eq():
     assert TargetPath.frame(1).region(".css") != TargetPath.frame(2).region(".css")
 
 
-def test_target_path_to_dict_region_by_css():
-    converted = TargetPath.region(".css").to_dict()
+def test_target_path_to_dict_region_by_css(obj_reg):
+    converted = TargetPath.region(".css").to_dict(obj_reg)
 
     assert converted == {"type": "css selector", "selector": ".css"}
 
 
-def test_target_path_to_dict_region_element():
-    converted = TargetPath.region(DummyElement("1")).to_dict()
+def test_target_path_to_dict_region_element(obj_reg):
+    converted = TargetPath.region(DummyElement("1")).to_dict(obj_reg)
 
     assert converted == {"elementId": "1"}
 
 
-def test_target_path_to_dict_shadow_by_css():
-    converted = TargetPath.shadow(".css").to_dict()
+def test_target_path_to_dict_shadow_by_css(obj_reg):
+    converted = TargetPath.shadow(".css").to_dict(obj_reg)
 
     assert converted == {"type": "css selector", "selector": ".css"}
 
 
-def test_target_path_to_dict_shadow_by_css_region_by_css():
-    converted = TargetPath.shadow("#s").region(".css").to_dict()
+def test_target_path_to_dict_shadow_by_css_region_by_css(obj_reg):
+    converted = TargetPath.shadow("#s").region(".css").to_dict(obj_reg)
 
     assert converted == {
         "type": "css selector",
@@ -276,8 +282,10 @@ def test_target_path_to_dict_shadow_by_css_region_by_css():
     }
 
 
-def test_target_path_to_dict_shadow_by_xpath_shadow_by_css_region_by_css():
-    converted = TargetPath.shadow(By.XPATH, "//x").shadow("#s").region(".css").to_dict()
+def test_target_path_to_dict_shadow_by_xpath_shadow_by_css_region_by_css(obj_reg):
+    converted = (
+        TargetPath.shadow(By.XPATH, "//x").shadow("#s").region(".css").to_dict(obj_reg)
+    )
 
     assert converted == {
         "type": "xpath",
@@ -293,8 +301,8 @@ def test_target_path_to_dict_shadow_by_xpath_shadow_by_css_region_by_css():
     }
 
 
-def test_target_path_to_dict_shadow_element_region_by_css():
-    converted = TargetPath.shadow(DummyElement(1)).region(".css").to_dict()
+def test_target_path_to_dict_shadow_element_region_by_css(obj_reg):
+    converted = TargetPath.shadow(DummyElement(1)).region(".css").to_dict(obj_reg)
 
     assert converted == {
         "elementId": 1,
@@ -302,8 +310,8 @@ def test_target_path_to_dict_shadow_element_region_by_css():
     }
 
 
-def test_target_path_to_dict_frame_by_css_region_by_css():
-    converted = TargetPath.frame(By.CSS_SELECTOR, "#s").region(".css").to_dict()
+def test_target_path_to_dict_frame_by_css_region_by_css(obj_reg):
+    converted = TargetPath.frame(By.CSS_SELECTOR, "#s").region(".css").to_dict(obj_reg)
 
     assert converted == {
         "type": "css selector",
@@ -312,25 +320,25 @@ def test_target_path_to_dict_frame_by_css_region_by_css():
     }
 
 
-def test_target_path_to_dict_region_by_id():
-    converted = TargetPath.region(By.ID, "id").to_dict()
+def test_target_path_to_dict_region_by_id(obj_reg):
+    converted = TargetPath.region(By.ID, "id").to_dict(obj_reg)
 
     assert converted == {"type": "id", "selector": "id"}
 
 
-def test_target_path_to_dict_region_by_tag_name():
-    converted = TargetPath.region(By.TAG_NAME, "tag").to_dict()
+def test_target_path_to_dict_region_by_tag_name(obj_reg):
+    converted = TargetPath.region(By.TAG_NAME, "tag").to_dict(obj_reg)
 
     assert converted == {"type": "tag name", "selector": "tag"}
 
 
-def test_target_path_to_dict_region_by_class_name():
-    converted = TargetPath.region(By.CLASS_NAME, "class").to_dict()
+def test_target_path_to_dict_region_by_class_name(obj_reg):
+    converted = TargetPath.region(By.CLASS_NAME, "class").to_dict(obj_reg)
 
     assert converted == {"type": "class name", "selector": "class"}
 
 
-def test_target_path_to_dict_region_by_name():
-    converted = TargetPath.region(By.NAME, "name").to_dict()
+def test_target_path_to_dict_region_by_name(obj_reg):
+    converted = TargetPath.region(By.NAME, "name").to_dict(obj_reg)
 
     assert converted == {"type": "name", "selector": "name"}